Rework folder structure
