+++ packages/twenty-sdk/src/cli/cli.ts
@@ -2,8 +2,7 @@
 import { inspect } from 'util';
 import chalk from 'chalk';
 import { Command, CommanderError } from 'commander';
-import { AppCommand } from '@/cli/commands/app.command';
-import { AuthCommand } from '@/cli/commands/auth.command';
+import { AppCommand, AuthCommand } from '@/cli/commands/app.command';
 import { ConfigService } from '@/cli/services/config.service';
 import packageJson from '../../package.json';
