+++ packages/twenty-sdk/src/cli/commands/app/app-generate.ts
