+++ packages/twenty-sdk/src/cli/commands/app/app-logs.ts
