+++ packages/twenty-sdk/src/cli/commands/app/app-publish.ts
@@ -1,0 +1,8 @@
+// Placeholder for app publish command
+// TODO: Implement application publishing functionality
+
+export class AppPublishCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
