+++ packages/twenty-sdk/src/cli/commands/app/app-sync.ts
