+++ packages/twenty-sdk/src/cli/commands/app/app-test.ts
@@ -1,0 +1,8 @@
+// Placeholder for app test command
+// TODO: Implement application testing functionality
+
+export class AppTestCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
