+++ packages/twenty-sdk/src/cli/commands/app/app-uninstall.ts
