+++ packages/twenty-sdk/src/cli/commands/app/index.ts
@@ -1,0 +1,9 @@
+export { AppWatchCommand } from './app-watch';
+export { AppBuildCommand, type BuildCommandOptions } from './app-build';
+export { AppSyncCommand } from './app-sync';
+export { AppUninstallCommand } from './app-uninstall';
+export { AppLogsCommand } from './app-logs';
+export { AppTestCommand } from './app-test';
+export { AppPublishCommand } from './app-publish';
+export { AppAddCommand, SyncableEntity, isSyncableEntity } from './app-add';
+export { AppGenerateCommand } from './app-generate';
