+++ packages/twenty-sdk/src/cli/commands/auth/auth-logout.ts
@@ -1,0 +1,24 @@
+import chalk from 'chalk';
+import { ConfigService } from '@/cli/services/config.service';
+
+export class AuthLogoutCommand {
+  private configService = new ConfigService();
+
+  async execute(): Promise<void> {
+    try {
+      await this.configService.clearConfig();
+      const activeWorkspace = ConfigService.getActiveWorkspace();
+      console.log(
+        chalk.green(
+          `✓ Successfully logged out (workspace: ${activeWorkspace})`,
+        ),
+      );
+    } catch (error) {
+      console.error(
+        chalk.red('Logout failed:'),
+        error instanceof Error ? error.message : error,
+      );
+      process.exit(1);
+    }
+  }
+}
