+++ packages/twenty-sdk/src/cli/commands/auth/auth-switch.ts
@@ -1,0 +1,8 @@
+// Placeholder for auth switch command
+// TODO: Implement workspace switching functionality
+
+export class AuthSwitchCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
