+++ packages/twenty-sdk/src/cli/commands/auth/index.ts
@@ -1,0 +1,4 @@
+export { AuthLoginCommand } from './auth-login';
+export { AuthLogoutCommand } from './auth-logout';
+export { AuthStatusCommand } from './auth-status';
+export { AuthSwitchCommand } from './auth-switch';
