+++ packages/twenty-sdk/src/cli/commands/entity/entity-add.ts
@@ -1,0 +1,8 @@
+// Placeholder for entity add command
+// TODO: Implement entity add functionality
+
+export class EntityAddCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
