+++ packages/twenty-sdk/src/cli/commands/entity/index.ts
@@ -1,0 +1,1 @@
+export { EntityAddCommand } from './entity-add';
