+++ packages/twenty-sdk/src/cli/commands/instance/index.ts
@@ -1,0 +1,3 @@
+export { InstanceSetupCommand } from './instance-setup';
+export { InstanceUpgradeCommand } from './instance-upgrade';
+export { InstanceRunCommand } from './instance-run';
