+++ packages/twenty-sdk/src/cli/commands/instance/instance-run.ts
@@ -1,0 +1,8 @@
+// Placeholder for instance run command
+// TODO: Implement instance run functionality
+
+export class InstanceRunCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
