+++ packages/twenty-sdk/src/cli/commands/instance/instance-setup.ts
@@ -1,0 +1,8 @@
+// Placeholder for instance setup command
+// TODO: Implement instance setup functionality
+
+export class InstanceSetupCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
