+++ packages/twenty-sdk/src/cli/commands/instance/instance-upgrade.ts
@@ -1,0 +1,8 @@
+// Placeholder for instance upgrade command
+// TODO: Implement instance upgrade functionality
+
+export class InstanceUpgradeCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
