+++ packages/twenty-sdk/src/cli/commands/sdk/index.ts
@@ -1,0 +1,1 @@
+export { SdkGenerateCommand } from './sdk-generate';
