+++ packages/twenty-sdk/src/cli/commands/sdk/sdk-generate.ts
@@ -1,0 +1,8 @@
+// Placeholder for sdk generate command
+// TODO: Implement SDK generation functionality
+
+export class SdkGenerateCommand {
+  async execute(): Promise<void> {
+    throw new Error('Not implemented');
+  }
+}
